[ExampleMod] Reorganize ModConfigShowcases
* Split ModConfigShowcases file into separate files (one file per each config section)

* Move custom data types from ModConfigShowcases to CustomDataTypes namespace (except Corner data type that is present in CornetElement class)

* Move custom UI (ConfigElements) from ModConfigShowcases to CustomDataTypes namespace
